backend: StarSystemQueries
